requirements-dev.txt
boto3
paramiko
python-decouple
